L1 hashing psw/backend/google-cloud-functions/vault/requirements.txt
# requirements.txt for Google Cloud Functions (Password Vault)
Flask==2.3.2
pycryptodome==3.20.0
google-cloud-firestore>=2.0.0